Fix panel/animation creation with large files

+++ jsolex-core/src/main/java/me/champeau/a4j/jsolex/processing/expr/impl/FfmegEncoder.java
@@ -15,6 +15,8 @@
  */
 package me.champeau.a4j.jsolex.processing.expr.impl;
 
+import me.champeau.a4j.jsolex.processing.event.ProgressEvent;
+import me.champeau.a4j.jsolex.processing.sun.Broadcaster;
 import me.champeau.a4j.jsolex.processing.sun.ImageUtils;
 import me.champeau.a4j.jsolex.processing.util.FileBackedImage;
 import me.champeau.a4j.jsolex.processing.util.ImageFormat;
@@ -26,13 +28,14 @@ import org.slf4j.LoggerFactory;
 import java.io.BufferedReader;
 import java.io.File;
 import java.io.IOException;
+import java.io.InputStreamReader;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
-import java.util.concurrent.Executors;
 import java.util.concurrent.TimeUnit;
+import java.util.concurrent.atomic.AtomicInteger;
+import java.util.stream.IntStream;
 
 public class FfmegEncoder {
     private static final Logger LOGGER = LoggerFactory.getLogger(FfmegEncoder.class);
@@ -41,31 +44,37 @@ public class FfmegEncoder {
         return Holder.isAvailable();
     }
 
-    public static void encode(List<Object> images, File outputFile, int msBetweenFrames) throws IOException {
+    public static boolean encode(Broadcaster broadcaster,
+                              List<Object> images,
+                              File outputFile,
+                              int msBetweenFrames) throws IOException {
         // first step is to export each image in a temporary directory
         // and each frame must be named with a sequence number, eg. frame-0001.png
         var tempDir = Files.createTempDirectory("jsolex-ffmpeg-");
-        var files = new ArrayList<File>();
+        List<File> frames = null;
         try {
-            try (var executor = Executors.newCachedThreadPool()) {
-                int sequenceNumber = 0;
-                for (Object image : images) {
-                    var fileName = String.format("frame-%04d.png", sequenceNumber++);
+            broadcaster.broadcast(ProgressEvent.of(0, "Exporting frames"));
+            var progress = new AtomicInteger(0);
+            frames = IntStream.range(0, images.size())
+                .parallel()
+                .mapToObj(i -> {
+                    var fileName = String.format("frame-%04d.png", i);
                     var file = new File(tempDir.toFile(), fileName);
-                    files.add(file);
-                    executor.submit(() -> {
-                        var img = image;
-                        if (img instanceof FileBackedImage fileBackedImage) {
-                            img = fileBackedImage.unwrapToMemory();
-                        }
-                        if (img instanceof ImageWrapper32 mono) {
-                            ImageUtils.writeMonoImage(mono.width(), mono.height(), mono.data(), file, Set.of(ImageFormat.PNG));
-                        } else if (img instanceof RGBImage rgb) {
-                            ImageUtils.writeRgbImage(rgb.width(), rgb.height(), rgb.r(), rgb.g(), rgb.b(), file, Set.of(ImageFormat.PNG));
-                        }
-                    });
-                }
-            }
+                    var img = images.get(i);
+                    if (img instanceof FileBackedImage fileBackedImage) {
+                        img = fileBackedImage.unwrapToMemory();
+                    }
+                    if (img instanceof ImageWrapper32 mono) {
+                        ImageUtils.writeMonoImage(mono.width(), mono.height(), mono.data(), file, Set.of(ImageFormat.PNG));
+                    } else if (img instanceof RGBImage rgb) {
+                        ImageUtils.writeRgbImage(rgb.width(), rgb.height(), rgb.r(), rgb.g(), rgb.b(), file, Set.of(ImageFormat.PNG));
+                    }
+                    broadcaster.broadcast(ProgressEvent.of(progress.incrementAndGet() / (double) images.size(), "Exporting frames"));
+                    return file;
+                })
+                .toList();
+            broadcaster.broadcast(ProgressEvent.of(1, "Exporting frames"));
+            broadcaster.broadcast(ProgressEvent.of(0, "Encoding (FFMPEG)"));
             int framesPerSecond = 1000 / msBetweenFrames;
             // now we can use ffmpeg to encode the video
             var ffmpeg = new ProcessBuilder("ffmpeg",
@@ -74,27 +83,38 @@ public class FfmegEncoder {
                 "-y",
                 "-crf", "10",
                 "-c:v", "libx264",
+                "-vf", "pad=ceil(iw/2)*2:ceil(ih/2)*2",
+                "-pix_fmt", "yuv420p",
                 outputFile.getAbsolutePath())
                 .redirectErrorStream(true)
                 .start();
-            try (var reader = new BufferedReader(new java.io.InputStreamReader(ffmpeg.getInputStream()))) {
+            var sb = new StringBuilder();
+            try (var reader = new BufferedReader(new InputStreamReader(ffmpeg.getInputStream()))) {
                 String line;
                 while ((line = reader.readLine()) != null) {
-                    LOGGER.debug(line);
+                    sb.append(line).append("\n");
                 }
             }
+            LOGGER.debug("FFMPEG output: {}", sb);
             ffmpeg.waitFor(10, TimeUnit.MINUTES);
+            broadcaster.broadcast(ProgressEvent.of(1, "Encoding (FFMPEG)"));
+            if (ffmpeg.exitValue() != 0) {
+                return false;
+            }
         } catch (InterruptedException e) {
             Thread.currentThread().interrupt();
             throw new IOException("Unable to encode video", e);
         } finally {
-            // delete all temporary files
-            for (File file : files) {
-                Path path = file.toPath();
-                Files.deleteIfExists(path);
+            if (frames != null) {
+                // delete all temporary files
+                for (File file : frames) {
+                    Path path = file.toPath();
+                    Files.deleteIfExists(path);
+                }
             }
             Files.deleteIfExists(tempDir);
         }
+        return true;
     }
 
     private static class Holder {
@@ -105,7 +125,7 @@ public class FfmegEncoder {
             try {
                 var ffmpeg = new ProcessBuilder("ffmpeg", "-h")
                     .start();
-                isAvailable = ffmpeg.waitFor(1, TimeUnit.SECONDS) && ffmpeg.exitValue() == 0;
+                isAvailable = ffmpeg.waitFor(5, TimeUnit.SECONDS) && ffmpeg.exitValue() == 0;
             } catch (IOException | InterruptedException e) {
                 isAvailable = false;
             }
